temporary work :  add settings to make all random sort or newest.  add l10n of init schedules.  fix the bug of share by copy that not correctly remember and delete.  rename the presentation. add setExistRow method to schedule to next refine the config page. next :  refine the config page.try find if have other bug.  try merge with the origin.make a release version.

+++ lib/widgets/collection/entry_set_action_delegate.dart
@@ -7,6 +7,7 @@ import 'package:aves/model/entry/extensions/favourites.dart';
 import 'package:aves/model/entry/extensions/metadata_edition.dart';
 import 'package:aves/model/entry/extensions/props.dart';
 import 'package:aves/model/favourites.dart';
+import 'package:aves/model/filters/album.dart';
 import 'package:aves/model/filters/filters.dart';
 import 'package:aves/model/metadata/date_modifier.dart';
 import 'package:aves/model/naming_pattern.dart';
@@ -22,6 +23,7 @@ import 'package:aves/services/common/image_op_events.dart';
 import 'package:aves/services/common/services.dart';
 import 'package:aves/theme/durations.dart';
 import 'package:aves/theme/themes.dart';
+import 'package:aves/utils/android_file_utils.dart';
 import 'package:aves/utils/collection_utils.dart';
 import 'package:aves/utils/mime_utils.dart';
 import 'package:aves/widgets/common/action_mixins/entry_editor.dart';
@@ -47,6 +49,8 @@ import 'package:flutter/scheduler.dart';
 import 'package:intl/intl.dart';
 import 'package:latlong2/latlong.dart';
 import 'package:provider/provider.dart';
+
+import 'collection_page.dart';
 
 class EntrySetActionDelegate with FeedbackMixin, PermissionAwareMixin, SizeAwareMixin, EntryEditorMixin, EntryStorageMixin {
   bool isVisible(
@@ -298,10 +302,11 @@ class EntrySetActionDelegate with FeedbackMixin, PermissionAwareMixin, SizeAware
     required BuildContext context,
     required Set<AvesEntry> entries,
     required bool enableBin,
+    bool isShareByCopyDelete = false,
   }) async {
     final pureTrash = entries.every((entry) => entry.trashed);
     if (enableBin && !pureTrash) {
-      await doMove(context, moveType: MoveType.toBin, entries: entries);
+      await doMove(context, moveType: MoveType.toBin, entries: entries,isShareByCopyDelete:isShareByCopyDelete);
       return;
     }
 
@@ -310,12 +315,15 @@ class EntrySetActionDelegate with FeedbackMixin, PermissionAwareMixin, SizeAware
     final storageDirs = entries.map((e) => e.storageDirectory).whereNotNull().toSet();
     final todoCount = entries.length;
 
-    if (!await showSkippableConfirmationDialog(
-      context: context,
-      type: ConfirmationDialog.deleteForever,
-      message: l10n.deleteEntriesConfirmationDialogMessage(todoCount),
-      confirmationButtonLabel: l10n.deleteButtonLabel,
-    )) return;
+    // in remove the share by copy delete, should not ask confirm.
+    if(!isShareByCopyDelete){
+      if (!await showSkippableConfirmationDialog(
+        context: context,
+        type: ConfirmationDialog.deleteForever,
+        message: l10n.deleteEntriesConfirmationDialogMessage(todoCount),
+        confirmationButtonLabel: l10n.deleteButtonLabel,
+      )) return;
+    }
 
     if (!await checkStoragePermissionForAlbums(context, storageDirs, entries: entries)) return;
 
@@ -398,6 +406,7 @@ class EntrySetActionDelegate with FeedbackMixin, PermissionAwareMixin, SizeAware
     Set<AvesEntry> todoItems,
     Future<Set<EntryDataType>> Function(AvesEntry entry) op, {
     bool showResult = true,
+    bool isShareByCopy = false,
   }) async {
     final selectionDirs = todoItems.map((e) => e.directory).whereNotNull().toSet();
     final todoCount = todoItems.length;
@@ -448,7 +457,6 @@ class EntrySetActionDelegate with FeedbackMixin, PermissionAwareMixin, SizeAware
         if (dataTypes.contains(EntryDataType.aspectRatio)) {
           source.onAspectRatioChanged();
         }
-
         if (showResult) {
           final l10n = context.l10n;
           final successCount = successOps.length;
@@ -457,7 +465,33 @@ class EntrySetActionDelegate with FeedbackMixin, PermissionAwareMixin, SizeAware
             showFeedback(context, FeedbackType.warn, l10n.collectionEditFailureFeedback(count));
           } else {
             final count = editedOps.length;
-            showFeedback(context, FeedbackType.info, l10n.collectionEditSuccessFeedback(count));
+            // for the move feedback will dismiss in 3seconds,
+            // add a edit feed back action to jump to the copied dir.
+            if(isShareByCopy){
+              final navigator = Navigator.maybeOf(context);
+              bool highlightTest(AvesEntry entry) => todoItems.contains(entry);
+              SnackBarAction action = SnackBarAction(
+                label: l10n.showButtonLabel,
+                onPressed: () {
+                  if (navigator != null) {
+                    unawaited(Navigator.maybeOf(context)?.pushAndRemoveUntil(
+                      MaterialPageRoute(
+                        settings: const RouteSettings(name: CollectionPage.routeName),
+                        builder: (context) => CollectionPage(
+                          source: source,
+                          filters: {AlbumFilter(androidFileUtils.avesShareByCopyPath,null)},
+                          highlightTest:highlightTest ,
+                        ),
+                      ),
+                          (route) => false,
+                    ));
+                  }
+                },
+              );
+              showFeedback(context, FeedbackType.info, l10n.collectionEditSuccessFeedback(count),action);
+            }else{
+              showFeedback(context, FeedbackType.info, l10n.collectionEditSuccessFeedback(count));
+            }
           }
         }
       },
@@ -520,7 +554,8 @@ class EntrySetActionDelegate with FeedbackMixin, PermissionAwareMixin, SizeAware
     await _edit(context, entries, (entry) => entry.flip());
   }
 
-  Future<void> setDateToNow(BuildContext context, {Set<AvesEntry>? entries, DateModifier? modifier, bool showResult = true,bool showConfirm = true,}) async {
+  Future<void> setDateToNow(BuildContext context, {Set<AvesEntry>? entries, DateModifier? modifier,
+    bool showResult = true,bool showConfirm = true,bool isShareByCopy = false}) async {
     entries ??= await _getEditableTargetItems(context, canEdit: (entry) => entry.canEditDate);
     if (entries == null || entries.isEmpty) return;
 
@@ -536,7 +571,8 @@ class EntrySetActionDelegate with FeedbackMixin, PermissionAwareMixin, SizeAware
 
     final modifier = DateModifier.setCustom(const {}, dateTime);
     if (modifier == null) return;
-    await _edit(context, entries, (entry) => entry.editDate(modifier!), showResult: showResult);
+    await _edit(context, entries, (entry) => entry.editDate(modifier!),
+        showResult: showResult,isShareByCopy: isShareByCopy);
   }
 
   Future<void> editDate(BuildContext context, {Set<AvesEntry>? entries, DateModifier? modifier, bool showResult = true}) async {
