+++ lib/widgets/common/action_mixins/entry_storage.dart
@@ -34,7 +34,6 @@ import 'package:aves/widgets/viewer/controls/notifications.dart';
 import 'package:aves_model/aves_model.dart';
 import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/scheduler.dart';
 import 'package:provider/provider.dart';
 
 import '../../../model/foreground_wallpaper/share_copied_entry.dart';
@@ -143,74 +142,34 @@ mixin EntryStorageMixin on FeedbackMixin, PermissionAwareMixin, SizeAwareMixin {
 
   Future<void> deleteExpiredShareCopied(BuildContext context) async {
     final source = context.read<CollectionSource>();
-    await shareCopiedEntries.init();
-    final todoEntries = source.allEntries.where((entry) =>
-        {AlbumFilter(androidFileUtils.avesShareByCopyPath,null)}.any((f) => f.test(entry))).toSet();
+    //await shareCopiedEntries.init();
+    final todoEntries = source.allEntries
+        .where((entry) => {AlbumFilter(androidFileUtils.avesShareByCopyPath, null)}.any((f) => f.test(entry)))
+        .toSet();
     final expiredEntries = todoEntries.where(shareCopiedEntries.isExpiredCopied).toSet();
-    if (expiredEntries.isEmpty) return ;
+    if (expiredEntries.isEmpty) return;
     debugPrint('$runtimeType deleteExpiredShareCopied\n'
         'todoEntries $todoEntries \n'
         'expiredEntries $expiredEntries\n');
 
     final entriesByDestination = <String, Set<AvesEntry>>{};
     entriesByDestination[AndroidFileUtils.trashDirPath] = expiredEntries;
-    // source.pauseMonitoring();
-    final destinationAlbums = entriesByDestination.keys.toSet();
+
     debugPrint('$runtimeType deleteExpiredShareCopied\n'
         'settings.enableBin ${settings.enableBin} \n'
-        ' settings.shareByCopyExpiredRemoveUseBin ${ settings.shareByCopyExpiredRemoveUseBin}\n');
-    if(settings.enableBin && settings.shareByCopyExpiredRemoveUseBin){
-      debugPrint('$runtimeType deleteExpiredShareCopied delete to bin \n');
-      final processed = <MoveOpEvent>{};
-      final completer = Completer<Set<String>>();
-      final opId = mediaEditService.newOpId;
-      mediaEditService.move(
-        opId: opId,
-        entriesByDestination: entriesByDestination,
-        copy: false,
-        // there should be no file conflict, as the target directory itself does not exist
-        nameConflictStrategy: NameConflictStrategy.rename,
-      ).listen(
-        processed.add,
-        onError: completer.completeError,
-        onDone: () async {
-          final successOps = processed.where((e) => e.success).toSet();
-
-          // move
-          final movedOps = successOps.where((v) => !v.skipped && !v.deleted).toSet();
-          await source.updateAfterMove(
-            todoEntries: expiredEntries,
-            moveType: MoveType.toBin,
-            destinationAlbums: destinationAlbums,
-            movedOps: movedOps,
-          );
-          // delete (when trying to move to bin obsolete entries)
-          final deletedOps = successOps.where((v) => v.deleted).toSet();
-          final deletedUris = deletedOps.map((event) => event.uri).toSet();
-          await source.removeEntries(deletedUris, includeTrash: true);
-          // source.resumeMonitoring();
-          completer.complete(deletedUris);
-        },
-      );
-      await completer.future;
-    }else{
-      debugPrint('$runtimeType deleteExpiredShareCopied delete to forvever \n');
-      final processed = <ImageOpEvent>{};
-      final completer = Completer<Set<String>>();
-      mediaEditService.delete(entries: expiredEntries).listen(
-        processed.add,
-        onError: completer.completeError,
-        onDone: () async {
-          final successOps = processed.where((e) => e.success).toSet();
-          final deletedOps = successOps.where((e) => !e.skipped).toSet();
-          final deletedUris = deletedOps.map((event) => event.uri).toSet();
-          // source.resumeMonitoring();
-          completer.complete(deletedUris);
-        },
-      );
-      await completer.future;
-    }
-   // return await completer.future;
+        ' settings.shareByCopyExpiredRemoveUseBin ${settings.shareByCopyExpiredRemoveUseBin}\n');
+    source.pauseMonitoring();
+    final useBin = settings.enableBin && settings.shareByCopyExpiredRemoveUseBin;
+    debugPrint('$runtimeType deleteExpiredShareCopied delete to useBin $useBin \n');
+    // delete forever trashed items
+    await EntrySetActionDelegate().doDelete(
+      context: context,
+      entries: expiredEntries,
+      enableBin: useBin,
+      isShareByCopyDelete: true,
+    );
+    await shareCopiedEntries.removeEntries(expiredEntries);
+    //await delegate.setDateToNow(context,entries:todoEntries,showConfirm: false);
     return;
   }
 
@@ -220,6 +179,7 @@ mixin EntryStorageMixin on FeedbackMixin, PermissionAwareMixin, SizeAwareMixin {
     required Map<String, Iterable<AvesEntry>> entriesByDestination,
     bool hideShowAction = false,
     VoidCallback? onSuccess,
+    bool isShareByCopyDelete = false,
   }) async {
     final entries = entriesByDestination.values.expand((v) => v).toSet();
     final todoCount = entries.length;
@@ -252,7 +212,8 @@ mixin EntryStorageMixin on FeedbackMixin, PermissionAwareMixin, SizeAwareMixin {
         if (await destinationDirectory.exists()) ...destinationDirectory.listSync().map((v) => pContext.basename(v.path)),
       ];
       final uniqueNames = names.toSet();
-      if (uniqueNames.length < names.length) {
+      //t4y: in share by copy, always rename the items.to avoid the show dialog break the operation.
+      if (uniqueNames.length < names.length && moveType != MoveType.shareByCopy) {
         final value = await showDialog<NameConflictStrategy>(
           context: context,
           builder: (context) => AvesSingleSelectionDialog<NameConflictStrategy>(
@@ -303,8 +264,8 @@ mixin EntryStorageMixin on FeedbackMixin, PermissionAwareMixin, SizeAwareMixin {
 
         source.resumeMonitoring();
 
-        // cleanup
-        if ({MoveType.move, MoveType.toBin}.contains(moveType)) {
+        // cleanup// t4y: should not delete the share by copy dir.
+        if ({MoveType.move, MoveType.toBin}.contains(moveType )&& !isShareByCopyDelete){
           await storageService.deleteEmptyRegularDirectories(originAlbums);
         }
 
@@ -327,19 +288,22 @@ mixin EntryStorageMixin on FeedbackMixin, PermissionAwareMixin, SizeAwareMixin {
             final navigator = Navigator.maybeOf(context);
             if (toBin) {
               if (movedEntries.isNotEmpty) {
-                action = SnackBarAction(
-                  label: Themes.asButtonLabel(l10n.entryActionRestore),
-                  onPressed: () {
-                    if (navigator != null) {
-                      doMove(
-                        navigator.context,
-                        moveType: MoveType.fromBin,
-                        entries: movedEntries,
-                        hideShowAction: true,
-                      );
-                    }
-                  },
-                );
+                // t4y: share by copy not need to restore.
+                if(!isShareByCopyDelete){
+                  action = SnackBarAction(
+                    label: Themes.asButtonLabel(l10n.entryActionRestore),
+                    onPressed: () {
+                      if (navigator != null) {
+                        doMove(
+                          navigator.context,
+                          moveType: MoveType.fromBin,
+                          entries: movedEntries,
+                          hideShowAction: true,
+                        );
+                      }
+                    },
+                  );
+                }
               }
             } else if (!hideShowAction) {
               action = SnackBarAction(
@@ -375,8 +339,9 @@ mixin EntryStorageMixin on FeedbackMixin, PermissionAwareMixin, SizeAwareMixin {
     required Set<AvesEntry> entries,
     bool hideShowAction = false,
     VoidCallback? onSuccess,
+    bool isShareByCopyDelete =false,
   }) async {
-    if (moveType == MoveType.toBin) {
+    if (moveType == MoveType.toBin && !isShareByCopyDelete) {
       final l10n = context.l10n;
       if (!await showSkippableConfirmationDialog(
         context: context,
@@ -384,6 +349,10 @@ mixin EntryStorageMixin on FeedbackMixin, PermissionAwareMixin, SizeAwareMixin {
         message: l10n.binEntriesConfirmationDialogMessage(entries.length),
         confirmationButtonLabel: l10n.deleteButtonLabel,
       )) return;
+    }
+    // t4y delete the expired items first.
+    if (moveType == MoveType.shareByCopy) {
+      await deleteExpiredShareCopied(context);
     }
 
     final entriesByDestination = <String, Set<AvesEntry>>{};
@@ -426,10 +395,16 @@ mixin EntryStorageMixin on FeedbackMixin, PermissionAwareMixin, SizeAwareMixin {
       moveType: moveType,
       entriesByDestination: entriesByDestination,
       onSuccess: onSuccess,
+      isShareByCopyDelete:isShareByCopyDelete,
     );
 
     if (moveType == MoveType.shareByCopy && settings.shareByCopyExpiredAutoRemove) {
-      await deleteExpiredShareCopied(context);
+      // await deleteExpiredShareCopied(context);
+      final source = context.read<CollectionSource>();
+      final todoEntries = source.allEntries.where((entry) =>
+          {AlbumFilter(androidFileUtils.avesShareByCopyPath,null)}.any((f) => f.test(entry)
+              && !entry.trashed)).where(shareCopiedEntries.isShareCopied).toSet();
+      await EntrySetActionDelegate().setDateToNow(context,entries:todoEntries,showConfirm: false,isShareByCopy: true);
     }
   }
